Simplify the CSongEditor class and update the main window accordingly
Several fields of the CSongEditor class have been removed (label,
...). The base class is now directly a CodeEditor: this removes the
need for a layout. The interface of the class is more Qt like (i.e.
without constructor parameter). A useless warning dialog has been
removed).

On the CMainWindow side, the code has been updated allowing to correct
a bug reported in the forum by Carreau with the modified status of the
different tabs. This change also remove duplicate editor. If the user
edits a currently edited song, the tab is only switched to the existing
tab.