Replace build-engine system by a more concise make-songbook-process
The new code is defined to let the user choose precisely what the make
process is going to do. There is no platform specific code, nor direct
links between the make process and the main application. The idea is
to be able to control everything from main-window where user preferences
can change the behavior (think about different compilation command)
of the tool.