handle make-songbook-process error message
This allows to display the LateX compilation logs widget when a LaTeX error occures (eg an unknown command in a song).
The problem was the the LaTeX process itself did not exit with a status error, thus this commit checks if the resulting pdf file to open actually exists and if not, emits an error signal.